refactor : domain계층 패키지 정리
