+++ aiku/aiku-common/src/main/java/common/domain/member/Member.java
@@ -1,7 +1,6 @@
 package common.domain.member;
 
 import common.domain.BaseTime;
-import common.domain.ServiceAgreement;
 import common.domain.Status;
 import common.domain.value_reference.TitleMemberValue;
 import jakarta.persistence.*;
