+++ aiku/aiku-common/src/main/java/common/domain/payment/PaymentProduct.java
@@ -1,6 +1,6 @@
-package common.domain;
+package common.domain.payment;
 
-import common.domain.member.Member;
+import common.domain.BaseTime;
 import common.domain.value_reference.MemberValue;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -14,7 +14,7 @@ import java.util.List;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class PaymentProduct extends BaseTime{
+public class PaymentProduct extends BaseTime {
 
     @Column(name = "paymentProductId")
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
