+++ aiku/aiku-common/src/main/java/common/domain/payment/PaymentProductType.java
@@ -1,4 +1,4 @@
-package common.domain;
+package common.domain.payment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
