+++ aiku/aiku-common/src/main/java/common/domain/shop/ShopProduct.java
@@ -1,5 +1,7 @@
-package common.domain;
+package common.domain.shop;
 
+import common.domain.BaseTime;
+import common.domain.Status;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -8,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class ShopProduct extends BaseTime{
+public class ShopProduct extends BaseTime {
 
     @Column(name = "shopProductId")
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
