+++ aiku/aiku-common/src/main/java/common/domain/term/AgreedType.java
@@ -1,4 +1,4 @@
-package common.domain;
+package common.domain.term;
 
 public enum AgreedType {
     MANDATORY, OPTIONAL;
