+++ aiku/aiku-common/src/main/java/common/domain/term/TermTitle.java
@@ -1,4 +1,4 @@
-package common.domain;
+package common.domain.term;
 
 public enum TermTitle {
     SERVICE, PERSONALINFO, LOCATION, MARKETING
