+++ aiku/aiku-main/src/main/java/aiku_main/repository/betting/BettingRepository.java
@@ -1,6 +1,6 @@
 package aiku_main.repository.betting;
 
-import common.domain.Betting;
+import common.domain.betting.Betting;
 import common.domain.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
