+++ aiku/aiku-main/src/main/java/aiku_main/repository/racing/RacingRepository.java
@@ -1,6 +1,6 @@
 package aiku_main.repository.racing;
 
-import common.domain.Racing;
+import common.domain.racing.Racing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RacingRepository extends JpaRepository<Racing, Long>, RacingRepositoryCustom {
