+++ aiku/aiku-main/src/main/java/aiku_main/service/term/TermService.java
@@ -3,8 +3,8 @@ package aiku_main.service.term;
 import aiku_main.dto.member.TermResDto;
 import aiku_main.exception.TermException;
 import aiku_main.repository.term.TermRepository;
-import common.domain.Term;
-import common.domain.TermTitle;
+import common.domain.term.Term;
+import common.domain.term.TermTitle;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
