+++ aiku/aiku-main/src/test/java/aiku_main/service/BettingServiceIntegrationTest.java
@@ -12,6 +12,7 @@ import aiku_main.service.betting.BettingService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.*;
+import common.domain.betting.Betting;
 import common.domain.member.Member;
 import common.domain.schedule.Schedule;
 import common.domain.schedule.ScheduleMember;
