+++ aiku/aiku-main/src/test/java/aiku_main/service/TeamServiceIntegrationTest.java
@@ -9,7 +9,9 @@ import aiku_main.service.team.TeamService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import common.domain.*;
+import common.domain.betting.Betting;
 import common.domain.member.Member;
+import common.domain.racing.Racing;
 import common.domain.schedule.Schedule;
 import common.domain.team.Team;
 import common.domain.team.TeamMember;
