+++ aiku/aiku-map/src/main/java/map/repository/RacingCommandRepository.java
@@ -1,8 +1,6 @@
 package map.repository;
 
-import common.domain.ExecStatus;
-import common.domain.Racing;
-import common.domain.Status;
+import common.domain.racing.Racing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RacingCommandRepository extends JpaRepository<Racing, Long>, RacingCommandRepositoryCustom {
