+++ aiku/aiku-map/src/main/java/map/repository/RacingQueryRepository.java
@@ -1,7 +1,7 @@
 package map.repository;
 
 import common.domain.ExecStatus;
-import common.domain.Racing;
+import common.domain.racing.Racing;
 import common.domain.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
