+++ aiku/aiku-map/src/test/java/map/service/RacingServiceTest.java
@@ -2,14 +2,13 @@ package map.service;
 
 import common.domain.ExecStatus;
 import common.domain.Location;
-import common.domain.Racing;
+import common.domain.racing.Racing;
 import common.domain.Status;
 import common.domain.member.Member;
 import common.domain.schedule.Schedule;
 import common.domain.schedule.ScheduleMember;
 import common.domain.team.Team;
 import common.domain.value_reference.TeamValue;
-import common.exception.PaidMemberLimitException;
 import jakarta.persistence.EntityManager;
 import map.dto.DataResDto;
 import map.dto.RacingAddDto;
