+++ aiku/aiku-payment/src/main/java/payment/repository/MemberRepository.java
@@ -1,11 +1,7 @@
 package payment.repository;
 
-import common.domain.PaymentProduct;
-import common.domain.PaymentProductType;
 import common.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 }
